requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.41.0
colorama==0.4.6
cron-descriptor==1.4.5
dj-database-url==2.3.0
Django==5.1.3
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-phonenumber-field==8.0.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
flower==2.0.1
gunicorn==23.0.0
humanize==4.11.0
idna==3.10
kombu==5.4.2
mock==5.1.0
packaging==24.2
phonenumbers==8.13.51
pillow==11.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2024.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
